refactor(auth): collapse oauth-client into @epicenter/auth/oauth-launchers
The oauth-client package was one file and one job that pairs 1:1 with
createOAuthAppAuth. Moving it under a subpath isolates oauth4webapi from
CLI bundles the same way /node already isolates Bun-only code, lets
auth-types own the canonical OAuthTokenGrant (was duplicated), and lets
browser/extension consumers import one auth package instead of two.

No PKCE, state, or token-exchange code rewritten. Server-side oauth
provider and resource-client verification untouched. The 14 oauth
ceremony tests move with the source and still pass.
