+++ apps/opensidian/package.json
@@ -23,7 +23,6 @@
 		"@epicenter/auth": "workspace:*",
 		"@epicenter/auth-svelte": "workspace:*",
 		"@epicenter/constants": "workspace:*",
-		"@epicenter/oauth-client": "workspace:*",
 		"@epicenter/skills": "workspace:*",
 		"@epicenter/svelte": "workspace:*",
 		"@tanstack/ai": "catalog:",
