+++ packages/auth/tsconfig.json
@@ -1,6 +1,7 @@
 {
 	"extends": "../../tsconfig.base.json",
 	"compilerOptions": {
+		"lib": ["ESNext", "DOM", "DOM.Iterable"],
 		"types": ["bun"],
 		"noUnusedLocals": true,
 		"noUnusedParameters": true
