--- packages/oauth-client/package.json
@@ -1,25 +1,0 @@
-{
-	"name": "@epicenter/oauth-client",
-	"version": "0.1.0",
-	"private": true,
-	"type": "module",
-	"description": "OAuth 2.1 PKCE launchers for hosted Epicenter auth.",
-	"main": "./src/index.ts",
-	"types": "./src/index.ts",
-	"exports": {
-		".": "./src/index.ts"
-	},
-	"scripts": {
-		"typecheck": "tsc --noEmit",
-		"test": "bun test"
-	},
-	"dependencies": {
-		"oauth4webapi": "catalog:",
-		"wellcrafted": "catalog:"
-	},
-	"devDependencies": {
-		"@types/bun": "catalog:",
-		"typescript": "catalog:"
-	},
-	"license": "MIT"
-}
